+++ qpmpc/mpc_qp.py
@@ -147,6 +147,10 @@ class MPCQP:
             self.q += mpc_problem.stage_state_cost_weight * np.dot(
                 c.T, self.Psi
             )
+        if mpc_problem.has_stage_input_target:
+            self.q -= mpc_problem.stage_input_cost_weight * (
+                mpc_problem.target_inputs
+            )
 
     def update_constraint_vector(self, mpc_problem: MPCProblem) -> None:
         """Update the inequality constraint vector.